requirements.txt
packaging
wheel
py-cpuinfo
transformers
torch
google-generativeai
python-dotenv
accelerate
bitsandbytes
peft
trl
sentencepiece
scipy
numpy
pandas
scikit-learn
jupyter
matplotlib
seaborn
huggingface_hub
datasets
evaluate
rouge_score
nltk
tqdm
wandb
deepspeed
optimum
vllm
